+++ packages/core/src/docdb/activation.ts
@@ -9,17 +9,18 @@ import { DBResourceNode } from './explorer/dbResourceNode'
 import { DocumentDBNode } from './explorer/docdbNode'
 import { DBClusterNode } from './explorer/dbClusterNode'
 import { DBInstanceNode } from './explorer/dbInstanceNode'
+import { addRegion } from './commands/addRegion'
 import { createCluster } from './commands/createCluster'
-import { createInstance } from './commands/createInstance'
 import { deleteCluster } from './commands/deleteCluster'
-import { deleteInstance } from './commands/deleteInstance'
 import { renameCluster } from './commands/renameCluster'
-import { renameInstance } from './commands/renameInstance'
+import { startCluster } from './commands/startCluster'
+import { stopCluster } from './commands/stopCluster'
+import { createInstance } from './commands/createInstance'
+import { deleteInstance } from './commands/deleteInstance'
 import { modifyInstance } from './commands/modifyInstance'
 import { rebootInstance } from './commands/rebootInstance'
-import { startCluster, stopCluster } from './commands/commands'
+import { renameInstance } from './commands/renameInstance'
 import { addTag, listTags, removeTag } from './commands/tagCommands'
-import { addRegion } from './commands/addRegion'
 
 /**
  * Activates DocumentDB components.
@@ -29,6 +30,10 @@ export async function activate(ctx: ExtContext): Promise<void> {
     ctx.extensionContext.subscriptions.push(
         Commands.register('aws.docdb.createCluster', async (node?: DocumentDBNode) => {
             await createCluster(node)
+        }),
+
+        Commands.register('aws.docdb.deleteCluster', async (node: DBClusterNode) => {
+            await deleteCluster(node)
         }),
 
         Commands.register('aws.docdb.renameCluster', async (node: DBClusterNode) => {
@@ -43,8 +48,8 @@ export async function activate(ctx: ExtContext): Promise<void> {
             await stopCluster(node)
         }),
 
-        Commands.register('aws.docdb.deleteCluster', async (node: DBClusterNode) => {
-            await deleteCluster(node)
+        Commands.register('aws.docdb.addRegion', async (node: DBClusterNode) => {
+            await addRegion(node)
         }),
 
         Commands.register('aws.docdb.createInstance', async (node: DBClusterNode) => {
@@ -55,16 +60,16 @@ export async function activate(ctx: ExtContext): Promise<void> {
             await deleteInstance(node)
         }),
 
-        Commands.register('aws.docdb.renameInstance', async (node: DBInstanceNode) => {
-            await renameInstance(node)
-        }),
-
         Commands.register('aws.docdb.modifyInstance', async (node: DBInstanceNode) => {
             await modifyInstance(node)
         }),
 
         Commands.register('aws.docdb.rebootInstance', async (node: DBInstanceNode) => {
             await rebootInstance(node)
+        }),
+
+        Commands.register('aws.docdb.renameInstance', async (node: DBInstanceNode) => {
+            await renameInstance(node)
         }),
 
         Commands.register('aws.docdb.listTags', async (node: DBResourceNode) => {
@@ -85,10 +90,6 @@ export async function activate(ctx: ExtContext): Promise<void> {
 
         Commands.register('aws.docdb.copyEndpoint', async (node?: DBResourceNode) => {
             await node?.copyEndpoint()
-        }),
-
-        Commands.register('aws.docdb.addRegion', async (node: DBClusterNode) => {
-            await addRegion(node)
         })
     )
 }
