+++ packages/core/src/docdb/commands/startCluster.ts
@@ -21,16 +21,3 @@ export function startCluster(node?: DBClusterNode): Promise<void> {
         }
     })
 }
-
-export function stopCluster(node?: DBClusterNode): Promise<void> {
-    return telemetry.docdb_stopCluster.run(async () => {
-        if (node?.arn && node?.regionCode) {
-            await node.client.stopCluster(node.arn)
-            getLogger().info('Stop cluster: %O', node.name)
-            void vscode.window.showInformationMessage(
-                localize('AWS.docdb.stopCluster.success', 'Stopping cluster: {0}', node.name)
-            )
-            node?.parent.refresh()
-        }
-    })
-}
